src/rarelink/cli/framework/framework.py
import typer

app = typer.Typer()

@app.command()
def install():
    """
    Install RareLink framework dependencies.
    """
    typer.echo("Installing RareLink framework dependencies...")
    # Example placeholder logic
    # You can include actual dependency installation logic here (e.g., pip install)
    typer.echo("Dependencies installed successfully.")

@app.command()
def update():
    """
    Update RareLink framework to the latest version.
    """
    typer.echo("Updating RareLink framework...")
    # Example placeholder logic
    # Include logic for updating the framework, like pulling updates
    typer.echo("Framework updated successfully.")
    
    
